gnome-common: Add recipe (from oe-core)
gnome-common is deprecated upstream (it's just a collection of
m4 macros so that's not a huge issue) and is soon not used in oe-core.

There are 6 recipes in meta-oe and 22 in meta-gnome that DEPEND on
gnome-common: Move gnome-common unchanged to meta-oe.

There is a hack still in autoconf-archive where some m4 files are
removed because they conflict with ones in gnome-common: this should
probably be switched over so the AX macros are removed from
gnome-common instead.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>